Add: test for add task function
