+++ utils/index.js
@@ -6,3 +6,5 @@ export const addTask = (task) => {
   localStorage.setItem('todos', task);
   document.body.innerHTML = '<div class="draggable" draggable="true"></div>';
 };
+
+// module.exports = { addTask };
